revision de codigo
cambio de nombre de algunas variables y eliminación de prints

+++ PrograDrones/src/Model/TripVariables.java
@@ -34,5 +34,5 @@ public class TripVariables {
     public int timeProx;
     public int cantDronesXPista;
   
-    public double cantDronesByIndividualTrip;
+    public int cantDronesByIndividualTrip;
 }
